major refactor on form composition
