+++ src/pages/ApiDroitsCnam.js
@@ -2,12 +2,18 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Form from '../components/Form';
 import Nav from '../components/Nav';
+import TextSection from '../components/form-sections/TextSection';
+import FranceconnectSection from '../components/form-sections/FranceconnectSection';
+import OrganisationSection from '../components/form-sections/OrganisationSection';
+import DescriptionSection from '../components/form-sections/DescriptionSection';
+import CadreJuridiqueSection from '../components/form-sections/CadreJuridiqueSection';
+import DonneesPersonnellesSection from '../components/form-sections/DonneesPersonnellesSection';
+import MiseEnOeuvreSection from '../components/form-sections/MiseEnOeuvreSection';
+import CguSection from '../components/form-sections/CguSection';
 
-// Description du contexte
 const target_api = 'api_droits_cnam';
 const title = "Demande d'accès à l'API Droits CNAM";
 
-// Le demandeur doit décrire le contexte d'usage de l'API
 const DemarcheDescription = () => (
   <div className="text-quote">
     <p>
@@ -33,16 +39,7 @@ const DemarcheDescription = () => (
     </p>
   </div>
 );
-const isFranceConnected = true;
 
-// Le demandeur doit séléctionner les données auxquelles il demande l'accès
-// Les perimètres de données ne sont pas disponibles
-// La ligne a été supprimée de la navigatrion {id: "#donnees", text: "Données"},
-const DonneesDescription = () => <React.Fragment />;
-const availableScopes = [];
-
-// Le demandeur valide les modalités d'utilisation
-const CguDescription = () => <React.Fragment />;
 const cguLink = '/docs/API_Droits_CNAM_CGU_20181210.pdf';
 
 const ApiDroitsCnam = ({
@@ -68,17 +65,16 @@ const ApiDroitsCnam = ({
       ]}
     />
     <div className="main">
-      <Form
-        enrollmentId={enrollmentId}
-        target_api={target_api}
-        title={title}
-        DemarcheDescription={DemarcheDescription}
-        isFranceConnected={isFranceConnected}
-        DonneesDescription={DonneesDescription}
-        availableScopes={availableScopes}
-        CguDescription={CguDescription}
-        cguLink={cguLink}
-      />
+      <Form enrollmentId={enrollmentId} target_api={target_api}>
+        <TextSection title={title} Description={DemarcheDescription} />
+        <FranceconnectSection />
+        <OrganisationSection />
+        <DescriptionSection />
+        <CadreJuridiqueSection />
+        <DonneesPersonnellesSection />
+        <MiseEnOeuvreSection />
+        <CguSection cguLink={cguLink} />
+      </Form>
     </div>
   </div>
 );
