+++ src/pages/ApiEntreprise.js
@@ -2,12 +2,18 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Form from '../components/Form';
 import Nav from '../components/Nav';
+import TextSection from '../components/form-sections/TextSection';
+import OrganisationSection from '../components/form-sections/OrganisationSection';
+import DescriptionSection from '../components/form-sections/DescriptionSection';
+import DonneesSection from '../components/form-sections/DonneesSection';
+import CadreJuridiqueSection from '../components/form-sections/CadreJuridiqueSection';
+import DonneesPersonnellesSection from '../components/form-sections/DonneesPersonnellesSection';
+import MiseEnOeuvreSection from '../components/form-sections/MiseEnOeuvreSection';
+import CguSection from '../components/form-sections/CguSection';
 
-// Description du contexte
 const target_api = 'api_entreprise';
 const title = "Demande d'accès à l'API Entreprise";
 
-// Le demandeur doit décrire le contexte d'usage de l'API
 const DemarcheDescription = () => (
   <div className="text-quote">
     <p>
@@ -33,7 +39,6 @@ const DemarcheDescription = () => (
     </p>
   </div>
 );
-const isFranceConnected = false;
 
 const contacts = {
   metier: {
@@ -60,9 +65,6 @@ const contacts = {
   },
 };
 
-// Le demandeur doit séléctionner les données auxquelles il demande l'accès
-// Les perimètres de données ne sont pas disponibles
-// La ligne a été supprimée de la navigatrion {id: "#donnees", text: "Données"},
 const DonneesDescription = () => (
   <div className="text-quote">
     <p>
@@ -167,8 +169,6 @@ const availableScopes = [
   },
 ];
 
-// Le demandeur valide les modalités d'utilisation
-const CguDescription = () => <React.Fragment />;
 const cguLink = 'https://entreprise.api.gouv.fr/cgu/';
 
 const ApiEntreprise = ({
@@ -198,18 +198,19 @@ const ApiEntreprise = ({
       ]}
     />
     <div className="main">
-      <Form
-        enrollmentId={enrollmentId}
-        target_api={target_api}
-        title={title}
-        DemarcheDescription={DemarcheDescription}
-        isFranceConnected={isFranceConnected}
-        contacts={contacts}
-        DonneesDescription={DonneesDescription}
-        availableScopes={availableScopes}
-        CguDescription={CguDescription}
-        cguLink={cguLink}
-      />
+      <Form enrollmentId={enrollmentId} target_api={target_api}>
+        <TextSection title={title} Description={DemarcheDescription} />
+        <OrganisationSection />
+        <DescriptionSection />
+        <DonneesSection
+          availableScopes={availableScopes}
+          DonneesDescription={DonneesDescription}
+        />
+        <CadreJuridiqueSection />
+        <DonneesPersonnellesSection />
+        <MiseEnOeuvreSection initialContacts={contacts} />
+        <CguSection cguLink={cguLink} />
+      </Form>
     </div>
   </div>
 );
